mcp_k8s_server/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.31.0
kubernetes==34.1.0
pyyaml==6.0.2
python-json-logger==2.0.4
httpx>=0.27.1
python-multipart>=0.0.9
pydantic>=2.11.0
requests>=2.31.0
